refactor: 팔레트 관련 API 분리
